Tetris video framework.
The first iteration of the tetris video framework. Draws the board,
the individual pieces, and does so within the game loop.

App is purposefully pub in Game. I don't want any App processing done
in Game and adding a setter/getter will encourage that.

EventQueue is included in this commit so WebApp isn't a useless
wrapper function. For benchmarking purposes.