Keyboard Events and Game Loop
Sends keydown events to the wasm app where they are added to an
event queue structure. Game loop reads the event queue and updates the
game accordingly.

Hack 0001 is included in this commit. The internal board structure is
larger than the rendered board and the bound offsets are hard coded
around this. The hack works well for bounds checking but makes rendering
unclean. Unsure of the correct fix.

Hack 0002 is included in this commit. When a piece is moved the internal
cur_piece structure is updated, the bounds are checked, and if the check
fails the piece is reverted to its previous state. This hack is not
thread-safe but I want a solution that does not clone cur_piece.